third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/triangular.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] triangular float32 2D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test triangular float32 actual 24.678754806518555 should be close enough to expected 0 by ULP distance: expected a number less than or equal to 0n but got 1103457815n
[FAIL] [required] triangular float16 2D tensor default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test triangular float16 actual 24.671875 should be close enough to expected 0 by ULP distance: expected a number less than or equal to 0 but got 20011
Harness: the test ran to completion.
